0.20200131: strip source
