+++ src/encode.h
