+++ src/huffman_fast_dec.h
@@ -3,10 +3,11 @@
 
 #include <stdint.h>
 
-typedef struct VLC {
-	int bits;
-	int16_t (*table)[2]; ///< code, bits
-	int table_size, table_allocated;
+typedef struct VLC
+{
+    int bits;
+    int16_t (*table)[2]; ///< code, bits
+    int table_size, table_allocated;
 } VLC;
 
 int build_vlc(VLC *vlc, const uint8_t *bits_table, const uint8_t *val_table, int nb_codes, int is_ac);
