+++ src/jpeg_optimizer.h
