--- std_quant_table.c
@@ -1,36 +1,0 @@
-#include "std_quant_table.h"
-
-const uint8_t std_quant_table[30][64] =
-{
-	{ 8, 4, 4, 4, 4, 4, 5, 5, 5, 5, 5, 5, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 7, 7, 7, 8, 8, 8, 7, 7, 7, 6, 6, 7, 7, 8, 8, 8, 8, 9, 9, 9, 8, 8, 8, 8, 9, 9, 10, 10, 10, 12, 12, 11, 11, 14, 14, 14, 17, 17, 20 },
-	{ 8, 6, 6, 7, 6, 7, 8, 8, 8, 8, 8, 8, 9, 9, 9, 10, 10, 10, 9, 9, 9, 9, 10, 10, 10, 10, 10, 10, 12, 12, 12, 10, 10, 10, 10, 10, 10, 10, 12, 12, 12, 12, 13, 14, 13, 13, 13, 12, 13, 14, 14, 15, 15, 15, 18, 18, 17, 17, 21, 21, 21, 25, 25, 31 },
-	{ 8, 8, 8, 9, 8, 9, 11, 11, 11, 11, 11, 11, 13, 12, 13, 13, 13, 13, 13, 13, 13, 13, 13, 13, 13, 14, 14, 14, 17, 17, 17, 14, 14, 14, 13, 13, 14, 14, 16, 16, 17, 17, 18, 19, 18, 17, 17, 17, 17, 19, 19, 20, 20, 20, 24, 24, 23, 23, 28, 28, 29, 34, 34, 41 },
-	{ 8, 10, 10, 11, 10, 11, 13, 13, 13, 13, 13, 13, 16, 15, 16, 16, 16, 16, 16, 16, 16, 16, 16, 16, 16, 18, 18, 18, 21, 21, 21, 18, 18, 18, 16, 16, 18, 18, 20, 20, 21, 21, 23, 23, 23, 21, 21, 21, 21, 23, 23, 25, 25, 25, 30, 30, 28, 28, 35, 35, 36, 43, 43, 51 },
-	{ 8, 12, 12, 14, 12, 14, 16, 16, 16, 16, 16, 16, 19, 18, 19, 20, 20, 20, 19, 19, 19, 19, 20, 20, 20, 21, 21, 21, 25, 25, 25, 21, 21, 21, 20, 20, 21, 21, 24, 24, 25, 25, 27, 28, 27, 26, 26, 25, 26, 28, 28, 30, 30, 30, 36, 36, 34, 34, 42, 42, 43, 51, 51, 62 },
-	{ 8, 14, 14, 16, 14, 16, 19, 19, 19, 19, 19, 19, 22, 21, 22, 23, 23, 23, 22, 22, 22, 22, 23, 23, 23, 25, 25, 25, 29, 29, 29, 25, 25, 25, 23, 23, 25, 25, 28, 28, 29, 29, 32, 33, 32, 30, 30, 29, 30, 33, 33, 35, 35, 35, 42, 42, 40, 40, 49, 49, 50, 60, 60, 72 },
-	{ 8, 16, 16, 19, 16, 19, 22, 22, 22, 22, 22, 22, 26, 24, 26, 27, 27, 27, 26, 26, 26, 26, 27, 27, 27, 29, 29, 29, 34, 34, 34, 29, 29, 29, 27, 27, 29, 29, 32, 32, 34, 34, 37, 38, 37, 35, 35, 34, 35, 38, 38, 40, 40, 40, 48, 48, 46, 46, 56, 56, 58, 69, 69, 83 },
-	{ 8, 18, 18, 21, 18, 21, 24, 24, 24, 24, 24, 24, 29, 27, 29, 30, 30, 30, 29, 29, 29, 29, 30, 30, 30, 32, 32, 32, 38, 38, 38, 32, 32, 32, 30, 30, 32, 32, 36, 36, 38, 38, 41, 42, 41, 39, 39, 38, 39, 42, 42, 45, 45, 45, 54, 54, 51, 51, 63, 63, 65, 77, 77, 93 },
-	{ 8, 20, 20, 23, 20, 23, 27, 27, 27, 27, 27, 27, 32, 30, 32, 33, 33, 33, 32, 32, 32, 32, 33, 33, 33, 36, 36, 36, 42, 42, 42, 36, 36, 36, 33, 33, 36, 36, 40, 40, 42, 42, 46, 47, 46, 43, 43, 42, 43, 47, 47, 50, 50, 50, 60, 60, 57, 57, 70, 70, 72, 86, 86, 103 },
-	{ 8, 22, 22, 26, 22, 26, 30, 30, 30, 30, 30, 30, 35, 33, 35, 37, 37, 37, 35, 35, 35, 35, 37, 37, 37, 39, 39, 39, 46, 46, 46, 39, 39, 39, 37, 37, 39, 39, 44, 44, 46, 46, 50, 52, 50, 48, 48, 46, 48, 52, 52, 55, 55, 55, 66, 66, 63, 63, 77, 77, 79, 94, 94, 114 },
-	{ 8, 24, 24, 28, 24, 28, 33, 33, 33, 33, 33, 33, 39, 36, 39, 40, 40, 40, 39, 39, 39, 39, 40, 40, 40, 43, 43, 43, 51, 51, 51, 43, 43, 43, 40, 40, 43, 43, 48, 48, 51, 51, 55, 57, 55, 52, 52, 51, 52, 57, 57, 60, 60, 60, 72, 72, 69, 69, 84, 84, 87, 103, 103, 124 },
-	{ 8, 26, 26, 30, 26, 30, 35, 35, 35, 35, 35, 35, 42, 39, 42, 43, 43, 43, 42, 42, 42, 42, 43, 43, 43, 47, 47, 47, 55, 55, 55, 47, 47, 47, 43, 43, 47, 47, 52, 52, 55, 55, 60, 61, 60, 56, 56, 55, 56, 61, 61, 65, 65, 65, 78, 78, 74, 74, 91, 91, 94, 112, 112, 134 },
-	{ 8, 28, 28, 33, 28, 33, 38, 38, 38, 38, 38, 38, 45, 42, 45, 47, 47, 47, 45, 45, 45, 45, 47, 47, 47, 50, 50, 50, 59, 59, 59, 50, 50, 50, 47, 47, 50, 50, 56, 56, 59, 59, 64, 66, 64, 61, 61, 59, 61, 66, 66, 70, 70, 70, 84, 84, 80, 80, 98, 98, 101, 120, 120, 145 },
-	{ 8, 30, 30, 35, 30, 35, 41, 41, 41, 41, 41, 41, 48, 45, 48, 50, 50, 50, 48, 48, 48, 48, 50, 50, 50, 54, 54, 54, 63, 63, 63, 54, 54, 54, 50, 50, 54, 54, 60, 60, 63, 63, 69, 71, 69, 65, 65, 63, 65, 71, 71, 75, 75, 75, 90, 90, 86, 86, 105, 105, 108, 129, 129, 155 },
-	{ 8, 32, 32, 38, 32, 38, 44, 44, 44, 44, 44, 44, 52, 48, 52, 54, 54, 54, 52, 52, 52, 52, 54, 54, 54, 58, 58, 58, 68, 68, 68, 58, 58, 58, 54, 54, 58, 58, 64, 64, 68, 68, 74, 76, 74, 70, 70, 68, 70, 76, 76, 80, 80, 80, 96, 96, 92, 92, 112, 112, 116, 138, 138, 166 },
-	{ 8, 34, 34, 40, 34, 40, 46, 46, 46, 46, 46, 46, 55, 51, 55, 57, 57, 57, 55, 55, 55, 55, 57, 57, 57, 61, 61, 61, 72, 72, 72, 61, 61, 61, 57, 57, 61, 61, 68, 68, 72, 72, 78, 80, 78, 74, 74, 72, 74, 80, 80, 85, 85, 85, 102, 102, 97, 97, 119, 119, 123, 146, 146, 176 },
-	{ 8, 36, 36, 42, 36, 42, 49, 49, 49, 49, 49, 49, 58, 54, 58, 60, 60, 60, 58, 58, 58, 58, 60, 60, 60, 65, 65, 65, 76, 76, 76, 65, 65, 65, 60, 60, 65, 65, 72, 72, 76, 76, 83, 85, 83, 78, 78, 76, 78, 85, 85, 90, 90, 90, 108, 108, 103, 103, 126, 126, 130, 155, 155, 186 },
-	{ 8, 38, 38, 45, 38, 45, 52, 52, 52, 52, 52, 52, 61, 57, 61, 64, 64, 64, 61, 61, 61, 61, 64, 64, 64, 68, 68, 68, 80, 80, 80, 68, 68, 68, 64, 64, 68, 68, 76, 76, 80, 80, 87, 90, 87, 83, 83, 80, 83, 90, 90, 95, 95, 95, 114, 114, 109, 109, 133, 133, 137, 163, 163, 197 },
-	{ 8, 40, 40, 47, 40, 47, 55, 55, 55, 55, 55, 55, 65, 60, 65, 67, 67, 67, 65, 65, 65, 65, 67, 67, 67, 72, 72, 72, 85, 85, 85, 72, 72, 72, 67, 67, 72, 72, 80, 80, 85, 85, 92, 95, 92, 87, 87, 85, 87, 95, 95, 100, 100, 100, 120, 120, 115, 115, 140, 140, 145, 172, 172, 207 },
-	{ 8, 42, 42, 49, 42, 49, 57, 57, 57, 57, 57, 57, 68, 63, 68, 70, 70, 70, 68, 68, 68, 68, 70, 70, 70, 76, 76, 76, 89, 89, 89, 76, 76, 76, 70, 70, 76, 76, 84, 84, 89, 89, 97, 99, 97, 91, 91, 89, 91, 99, 99, 105, 105, 105, 126, 126, 120, 120, 147, 147, 152, 181, 181, 217 },
-	{ 8, 44, 44, 52, 44, 52, 60, 60, 60, 60, 60, 60, 71, 66, 71, 74, 74, 74, 71, 71, 71, 71, 74, 74, 74, 79, 79, 79, 93, 93, 93, 79, 79, 79, 74, 74, 79, 79, 88, 88, 93, 93, 101, 104, 101, 96, 96, 93, 96, 104, 104, 110, 110, 110, 132, 132, 126, 126, 154, 154, 159, 189, 189, 228 },
-	{ 8, 46, 46, 54, 46, 54, 63, 63, 63, 63, 63, 63, 74, 69, 74, 77, 77, 77, 74, 74, 74, 74, 77, 77, 77, 83, 83, 83, 97, 97, 97, 83, 83, 83, 77, 77, 83, 83, 92, 92, 97, 97, 106, 109, 106, 100, 100, 97, 100, 109, 109, 115, 115, 115, 138, 138, 132, 132, 161, 161, 166, 198, 198, 238 },
-	{ 8, 48, 48, 57, 48, 57, 66, 66, 66, 66, 66, 66, 78, 72, 78, 81, 81, 81, 78, 78, 78, 78, 81, 81, 81, 87, 87, 87, 102, 102, 102, 87, 87, 87, 81, 81, 87, 87, 96, 96, 102, 102, 111, 114, 111, 105, 105, 102, 105, 114, 114, 120, 120, 120, 144, 144, 138, 138, 168, 168, 174, 207, 207, 249 },
-	{ 8, 50, 50, 59, 50, 59, 68, 68, 68, 68, 68, 68, 81, 75, 81, 84, 84, 84, 81, 81, 81, 81, 84, 84, 84, 90, 90, 90, 106, 106, 106, 90, 90, 90, 84, 84, 90, 90, 100, 100, 106, 106, 115, 118, 115, 109, 109, 106, 109, 118, 118, 125, 125, 125, 150, 150, 143, 143, 175, 175, 181, 215, 215, 255 },
-	{ 8, 52, 52, 61, 52, 61, 71, 71, 71, 71, 71, 71, 84, 78, 84, 87, 87, 87, 84, 84, 84, 84, 87, 87, 87, 94, 94, 94, 110, 110, 110, 94, 94, 94, 87, 87, 94, 94, 104, 104, 110, 110, 120, 123, 120, 113, 113, 110, 113, 123, 123, 130, 130, 130, 156, 156, 149, 149, 182, 182, 188, 224, 224, 255 },
-	{ 8, 54, 54, 64, 54, 64, 74, 74, 74, 74, 74, 74, 87, 81, 87, 91, 91, 91, 87, 87, 87, 87, 91, 91, 91, 97, 97, 97, 114, 114, 114, 97, 97, 97, 91, 91, 97, 97, 108, 108, 114, 114, 124, 128, 124, 118, 118, 114, 118, 128, 128, 135, 135, 135, 162, 162, 155, 155, 189, 189, 195, 232, 232, 255 },
-	{ 8, 56, 56, 66, 56, 66, 77, 77, 77, 77, 77, 77, 91, 84, 91, 94, 94, 94, 91, 91, 91, 91, 94, 94, 94, 101, 101, 101, 119, 119, 119, 101, 101, 101, 94, 94, 101, 101, 112, 112, 119, 119, 129, 133, 129, 122, 122, 119, 122, 133, 133, 140, 140, 140, 168, 168, 161, 161, 196, 196, 203, 241, 241, 255 },
-	{ 8, 58, 58, 68, 58, 68, 79, 79, 79, 79, 79, 79, 94, 87, 94, 97, 97, 97, 94, 94, 94, 94, 97, 97, 97, 105, 105, 105, 123, 123, 123, 105, 105, 105, 97, 97, 105, 105, 116, 116, 123, 123, 134, 137, 134, 126, 126, 123, 126, 137, 137, 145, 145, 145, 174, 174, 166, 166, 203, 203, 210, 250, 250, 255 },
-	{ 8, 60, 60, 71, 60, 71, 82, 82, 82, 82, 82, 82, 97, 90, 97, 101, 101, 101, 97, 97, 97, 97, 101, 101, 101, 108, 108, 108, 127, 127, 127, 108, 108, 108, 101, 101, 108, 108, 120, 120, 127, 127, 138, 142, 138, 131, 131, 127, 131, 142, 142, 150, 150, 150, 180, 180, 172, 172, 210, 210, 217, 255, 255, 255 },
-	{ 8, 62, 62, 73, 62, 73, 85, 85, 85, 85, 85, 85, 100, 93, 100, 104, 104, 104, 100, 100, 100, 100, 104, 104, 104, 112, 112, 112, 131, 131, 131, 112, 112, 112, 104, 104, 112, 112, 124, 124, 131, 131, 143, 147, 143, 135, 135, 131, 135, 147, 147, 155, 155, 155, 186, 186, 178, 178, 217, 217, 224, 255, 255, 255 },
-};
-
